+++ apps/web/src/services/uploader/index.ts
@@ -1,2 +1,0 @@
-export { default as useFileUploader } from "./useFileUploader";
-export { default as FileWithProgress } from "./FileWithProgress";